refactor: move core utilities to src/utils
