+++ tsconfig.json
@@ -1,6 +1,9 @@
 {
     "compilerOptions": {
         "baseUrl": ".",
+        "paths": {
+            "src/*": ["src/*"]
+        },
         "inlineSourceMap": true,
         "inlineSources": true,
         "module": "ESNext",
@@ -15,5 +18,5 @@
         "skipLibCheck": true,
         "lib": ["DOM", "ES5", "ES6", "ES7"]
     },
-    "include": ["**/*.ts"]
+    "include": ["src/**/*.ts"]
 }
